@@ -3,9 +3,9 @@ import 'dart:io';
 import 'package:flutter_stetho/src/http_client.dart';
 
 class StethoHttpOverrides extends HttpOverrides {
-  final String Function(Uri url, Map<String, String> environment)
+  final String Function(Uri url, Map<String, String> environment)?
       findProxyFromEnvironmentFn;
-  final HttpClient Function(SecurityContext context) createHttpClientFn;
+  final HttpClient Function(SecurityContext context)? createHttpClientFn;
 
   StethoHttpOverrides({
     this.findProxyFromEnvironmentFn,
@@ -13,22 +13,22 @@ class StethoHttpOverrides extends HttpOverrides {
   });
 
   @override
-  HttpClient createHttpClient(SecurityContext context) {
+  HttpClient createHttpClient(SecurityContext? context) {
     final client = createHttpClientFn != null
-        ? createHttpClientFn(context)
+        ? createHttpClientFn!(context!)
         : super.createHttpClient(context);
 
     if (Platform.isAndroid) {
-      return new StethoHttpClient(client);
+      return StethoHttpClient(client: client);
     }
 
     return client;
   }
 
   @override
-  String findProxyFromEnvironment(Uri url, Map<String, String> environment) {
+  String findProxyFromEnvironment(Uri? url, Map<String, String>? environment) {
     return findProxyFromEnvironmentFn != null
-        ? findProxyFromEnvironmentFn(url, environment)
-        : super.findProxyFromEnvironment(url, environment);
+        ? findProxyFromEnvironmentFn!(url!, environment!)
+        : super.findProxyFromEnvironment(url!, environment);
   }
 }
