@@ -1,14 +1,13 @@
 import 'dart:async';
 import 'dart:io';
 import 'dart:math';
-import 'dart:typed_data';
 
 import 'package:flutter_stetho/src/method_channel_controller.dart';
 
 /// Create a response transformer that can intercept and pipe the http response
 /// data to the Method channel
 StreamTransformer<List<int>, List<int>> createResponseTransformer(String id) {
-  return new StreamTransformer.fromHandlers(handleData: (data, sink) {
+  return StreamTransformer.fromHandlers(handleData: (data, sink) {
     sink.add(data);
     MethodChannelController.onDataReceived({"data": data, "id": id});
   }, handleError: (error, stacktrace, sink) {
@@ -46,7 +45,7 @@ Map<String, String> headersToMap(HttpHeaders headers) {
 ///
 ///     final String id = new Uuid().generateV4();
 class Uuid {
-  final Random _random = new Random();
+  final Random _random = Random();
 
   /// Generate a version 4 (random) uuid. This is a uuid scheme that only uses
   /// random numbers as the source of the generated uuid.
